src/requirements.txt
llfuse==1.5.0
pandas==2.1.3
patool==1.15.0
PyPAKParser==1.2.0
